requirements.txt
aiohttp==3.7.4.post0
aiosignal==1.2.0
altgraph==0.17.2
APScheduler==3.8.1
asgiref==3.4.1
async-generator==1.10
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==21.2.0
base58==2.1.1
bitarray==1.2.2
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
configparser==5.2.0
crayons==0.4.0
cryptography==36.0.1
cytoolz==0.11.2
dataclasses==0.6
DateTime==4.3
discord==1.7.3
discord.py==1.7.3
et-xmlfile==1.1.0
eth-abi==2.1.1
eth-account==0.5.6
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
Flask==2.0.2
frozenlist==1.2.0
future==0.18.2
h11==0.12.0
hexbytes==0.2.2
idna==3.3
iniconfig==1.1.1
ipfshttpclient==0.8.0a2
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==3.2.0
kaleido==0.2.1
lru-dict==1.1.7
MarkupSafe==2.0.1
multiaddr==0.0.9
multidict==5.2.0
netaddr==0.8.0
numpy==1.21.4
openpyxl==3.0.9
outcome==1.1.0
packaging==21.3
pandas==1.3.4
parsimonious==0.8.1
pefile==2021.9.3
Pillow==8.4.0
plotly==5.5.0
pluggy==1.0.0
protobuf==3.19.1
py==1.11.0
pycparser==2.21
pycryptodome==3.12.0
pycryptodomex==3.12.0
pyinstaller==4.8
pyinstaller-hooks-contrib==2022.0
pyOpenSSL==21.0.0
pyparsing==3.0.6
pyrsistent==0.18.0
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.2
python-minifier==2.5.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywin32==302
pywin32-ctypes==0.2.0
qrcode==7.3.1
requests==2.26.0
rlp==2.0.1
schedule==1.1.0
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
tabulate==0.8.9
tdqm==0.0.1
tenacity==8.0.1
toml==0.10.2
toolz==0.11.2
tqdm==4.62.3
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
varint==1.0.2
web3==5.25.0
webdriver-manager==3.5.2
websockets==9.1
Werkzeug==2.0.2
wsproto==1.0.0
XlsxWriter==3.0.2
yarl==1.7.2
zope.interface==5.4.0
